Set up a socket.io server
This is an app that benefits from real-time communication, since it's
fun for voters to see their own actions and those of others immediately
when they occur. For that reason, we're going to use WebSockets to
communicate. More specifically, we're going to use the Socket.io library
that provides a nice abstraction for WebSockets that works across
browsers. It also has a number of fallback mechanisms for clients that
don't support WebSockets.